refactor: Reorganize test files into proper folder structure
- Move all test files from root to tests/ subfolders
- Create tests/unit/ for unit tests (filename cleanup, training)
- Create tests/integration/ for integration tests (cross-entry, HTML formatting)
- Create tests/data/ for test data files (.srt, .progress)
- Create tests/debug/ for debug and analysis scripts
- Update test scripts to use plugin system instead of direct clients
- Fix import paths for moved test files
- Add pytest.ini for proper test configuration
- Add tests/README.md documenting the test structure
- All tests now compatible with plugin system architecture
