+++ pytest.ini
@@ -1,0 +1,4 @@
+[tool:pytest]
+pythonpath = .
+testpaths = tests
+addopts = -v
