--- test_fix.srt
@@ -1,29 +1,0 @@
-1
-00:00:55,000 --> 00:00:57,000
-This is a regular line.
-
-2
-00:00:57,819 --> 00:00:59,795
-(INDISTINCT RADIO CHATTER)
-
-3
-00:00:59,819 --> 00:01:01,825
-- What do we got?
-- Vic is Nia Fox.
-
-4
-00:01:02,000 --> 00:01:04,000
-Another normal sentence.
-
-5
-00:01:04,500 --> 00:01:06,000
-[DOOR SLAMS]
-
-6
-00:01:06,500 --> 00:01:08,000
-- This is incomplete dialogue
-that continues here.
-
-7
-00:01:08,500 --> 00:01:10,000
-Complete dialogue here.
