+++ tests/debug/debug_batch.py
