+++ tests/debug/debug_batch_detailed.py
